+++ src/modules/Language_Switcher/Switchers/Dropdown.php
@@ -7,7 +7,7 @@ namespace WP_Syntex\Polylang\Language_Switcher\Switchers;
 
 use PLL_Language;
 use WP_Syntex\Polylang\Language_Switcher\Assets;
-use WP_Syntex\Polylang\Language_Switcher\Switchers\Elements\Nav as Element;
+use WP_Syntex\Polylang\Language_Switcher\Switchers\Element\Nav as Element;
 
 defined( 'ABSPATH' ) || exit;
